ignore origin package files

+++ packager.sh
@@ -5,6 +5,9 @@ PKG_DIR=oq-python3.5-3.5.4
 BASE_URL="https://launchpad.net/debian/+archive/primary/+files"
 
 for f in python3.5_3.5.4-4.dsc python3.5_3.5.4.orig.tar.gz python3.5_3.5.4-4.debian.tar.xz; do
+    if [ -f "$f" ]; then
+        continue
+    fi
     curl -o "$f" -L "${BASE_URL}/${f}"
 done
